Allow psych v4 (#98)
In psych v4, the YAML.load method has been updated to provide a similar
interface and semantics to YAML.safe_load. Refactor the code to use
YAML.safe_load for a consistent behaviour between psych V3 and V4.

Psych will automatically convert strings that look like dates to Ruby
Date objects. We need to permit this class when loading CloudFormation
YAML.

The permitted_classes keyword argument for YAML.safe_load was introduced
in psych v3.1.0. So this is now the minimum compatible version of psych.